+++ src/libopentld/tld/TLD.cpp
@@ -623,7 +623,7 @@ void TLD::readFromFile(const char *path)
 
     int size = 2 * 2 * ec->numFeatures * ec->numTrees;
     ec->features = new float[size];
-    ec->numIndices = pow(2.0f, ec->numFeatures);
+    ec->numIndices = static_cast<int>(pow(2.0f, ec->numFeatures));
     ec->initPosteriors();
 
     for(int i = 0; i < ec->numTrees; i++)
